Split PubSubTemplate into Publisher and Subscriber
Changes PubSubTemplate to contain instances of PubSubPublisherTemplate
and PubSubSubscriberTemplate and delegates all method calls to them.

Fixes #816.
